refactor(runtime): split runtime builder
